View to show all inbound sms numbers for the platform admin user.
