+++ app/notify_client/inbound_number_client.py
@@ -12,7 +12,7 @@ class InboundNumberClient(NotifyAdminAPIClient):
         self.api_key = app.config['ADMIN_CLIENT_SECRET']
 
     def get_all_inbound_sms_number_service(self):
-        return self.get('/inbound_number')
+        return self.get('/inbound-number')
 
     def get_inbound_sms_number_for_service(self, service_id):
         return self.get('/inbound-number/service/{}'.format(service_id))
