+++ app/templates/views/platform-admin/_base_template.html
@@ -17,7 +17,8 @@
           ('Trial mode services', url_for('main.trial_services')),
           ('Providers', url_for('main.view_providers')),
           ('Organisations', url_for('main.organisations')),
-          ('Letter jobs', url_for('main.letter_jobs'))
+          ('Letter jobs', url_for('main.letter_jobs')),
+          ('Inbound SMS numbers', url_for('main.inbound_sms_admin'))
         ] %}
           <li>
             <a href="{{ url }}">
